requirements.txt
pandas
sklearn
tensorflow
keras
flask
numpy
scipy
pillow
matplotlib
h5py